Suranta_Fly/main_server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
aiohttp==3.9.1
pytest==7.4.3
httpx==0.25.2
alembic==1.12.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
tenacity==8.2.3
structlog==23.2.0
python-json-logger==2.0.7
cryptography==41.0.5
jinja2==3.1.2
psutil==5.9.6 